Use non-repeating timer for autocommitting
I was seeing some duplicate behavior when using the repeating timer.

Now, after making a change to the content of a note, schedule a call to
commit after 10 seconds. Also, when about to commit, invalidate all
other timers so we don't try to commit N times, where N is the number of
changes made to the text content (one change for each
addition/removal/etc)