Get push working without manually editing simulator files
This gets push behavior working by using a fork of a fork of SwiftGit2.

Before this, pushes were failing because the local repo had no push
refspecs configured.

Not sure if this is the right fix or not, but the fork adds a push
refspec for the master branch right after cloning.

So now when trying to auto-push, the matching refspec is found and used
to push changes.

NOTE: doesn't seem like this would work with a repo whose default branch
isn't "master"

See
https://github.com/Maaimusic/SwiftGit2-Public-Fork/compare/Maaimusic:2e4199b...iancanderson:e24a0c0